example_transform/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "example_transform"
version = "0.0.1"
dependencies = [
    "build == 0.10.0",
    "opencensus-ext-azure==1.1.9",
    "opencensus-ext-logging==0.1.1",
    "pandas == 1.5.3",
    "pandas-stubs == 1.5.3.230203",
    "pyspark == 3.4.1",
    "pytest == 7.2.1",
    "mypy == 1.0.0"
]

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = "testcontainers.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "opencensus.*"
ignore_missing_imports = true
